+++ src/main/java/com/searching/instagram/dto/SaveDTO.java
@@ -22,10 +22,10 @@ public class SaveDTO {
     private Long id;
 
     @Positive
-    private Integer profileId;
+    private Long profileId;
 
     @Positive
-    private Integer postId;
+    private Long postId;
 
     private SaveType saveType;
